requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Babel==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
requests==2.31.0
SQLAlchemy==2.0.21
Flask-Limiter==3.5.0
redis==5.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3